Refactor forms project
