--- EngineImplForms/FormCanvas.cs
@@ -1,17 +1,0 @@
-namespace EngineImplForms
-{
-    public partial class FormCanvas : Form
-    {
-        Graphics graphics;
-
-        public FormCanvas()
-        {
-            InitializeComponent();
-        }
-
-        private void FormCanvas_Load(object sender, EventArgs e)
-        {
-            graphics = pictureBoxCanvas.CreateGraphics();
-        }
-    }
-}
