+++ RaycasterEngine.sln
@@ -9,14 +9,14 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EngineImpl", "EngineImpl\En
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Engine.Tests", "Engine.Tests\Engine.Tests.csproj", "{F373A983-C990-45F0-9D62-9EF774EE7D33}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EngineImplForms", "EngineImplForms\EngineImplForms.csproj", "{2DF2BE7E-2715-4772-A56A-F226CD0D5CA0}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{9061EA6A-1661-4137-BD6B-17179BB407BC}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EngineImpl.Terminal", "EngineImpl.Terminal\EngineImpl.Terminal.csproj", "{EC34AA5E-AE9D-4C5C-B661-2861F982965F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EngineImpl.Forms", "EngineImpl.Forms\EngineImpl.Forms.csproj", "{F2CACEA1-A3D7-4BBA-82EF-6F8EDD5A6BF1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -36,14 +36,14 @@ Global
 		{F373A983-C990-45F0-9D62-9EF774EE7D33}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F373A983-C990-45F0-9D62-9EF774EE7D33}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F373A983-C990-45F0-9D62-9EF774EE7D33}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2DF2BE7E-2715-4772-A56A-F226CD0D5CA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2DF2BE7E-2715-4772-A56A-F226CD0D5CA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2DF2BE7E-2715-4772-A56A-F226CD0D5CA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2DF2BE7E-2715-4772-A56A-F226CD0D5CA0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EC34AA5E-AE9D-4C5C-B661-2861F982965F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EC34AA5E-AE9D-4C5C-B661-2861F982965F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EC34AA5E-AE9D-4C5C-B661-2861F982965F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EC34AA5E-AE9D-4C5C-B661-2861F982965F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F2CACEA1-A3D7-4BBA-82EF-6F8EDD5A6BF1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F2CACEA1-A3D7-4BBA-82EF-6F8EDD5A6BF1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F2CACEA1-A3D7-4BBA-82EF-6F8EDD5A6BF1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F2CACEA1-A3D7-4BBA-82EF-6F8EDD5A6BF1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
